% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dfa.R
\name{compile_stan_model}
\alias{compile_stan_model}
\title{Compile Stan model}
\usage{
compile_stan_model(model_name)
}
\arguments{
\item{model_name}{String that is dfa, regime_1, hmm_gaussian, or corr}
}
\value{
Stan model.
}
\description{
Compile Stan model
}
