% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converge_rhat.R
\name{converge_rhat}
\alias{converge_rhat}
\title{Summarize Rhat convergence statistics across parameters}
\usage{
converge_rhat(fitted_model, threshold = 1.05)
}
\arguments{
\item{fitted_model}{A matrix of data to fit. Columns represent time element.}

\item{threshold}{Threshold for maximum Rhat, default = 1.05}
}
\description{
Pass in rstanfit model object, and optional threshold value for convergence. Returns boolean
}

