% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_regimes.R
\name{find_regimes}
\alias{find_regimes}
\title{Fit models with differing numbers of regimes to trend data}
\usage{
find_regimes(y, lower = 1, upper = 5, mcmc_iter = 2000, mcmc_chains = 3,
  threshold = 1.05)
}
\arguments{
\item{y}{Data, time series or trend from fitted DFA model.}

\item{lower}{Lower bound of number of regimes to evaluate, defaults to 1}

\item{upper}{Upper bound of number of regimes to evaluate, defaults to 5}

\item{mcmc_iter}{MCMC iterations, defaults to 2000}

\item{mcmc_chains}{MCMC chains, defaults to 3}

\item{threshold}{Threshold for maximum Rhat, default = 1.05}
}
\description{
Pass in rstanfit model object, and optional threshold value for
convergence. Returns boolean.
}
\examples{
\dontrun{
data(Nile)
find_regimes(y=log(Nile), upper=3)
}
}
