% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_regimes2.R
\name{find_regimes2}
\alias{find_regimes2}
\title{Fit models with differing numbers of regimes to trend data}
\usage{
find_regimes2(y, n_regimes = 2, iter = 2000, chains = 1, ...)
}
\arguments{
\item{y}{Data, time series or trend from fitted DFA model.}

\item{n_regimes}{Number of regimes to evaluate}

\item{iter}{MCMC iterations, defaults to 2000}

\item{chains}{MCMC chains, defaults to 1 (note that running multiple chains
may result in a label switching problem where the regimes are identified
with different IDs across chains).}

\item{...}{Other parameters to pass to \code{\link[rstan]{sampling}}}
}
\description{
Fit models with differing numbers of regimes to trend data
}
\examples{
\dontrun{
data(Nile)
find_regimes2(log(Nile))
}
}
