% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_swans.R
\name{find_swans}
\alias{find_swans}
\title{Find outlying "black swan" jumps in trends}
\usage{
find_swans(rotated_modelfit, threshold = 0.01)
}
\arguments{
\item{rotated_modelfit}{Output from \code{\link{rotate_trends}}}

\item{threshold}{A probability threshold below which to 
flag trend events as extreme}
}
\description{
Find outlying "black swan" jumps in trends
}
\examples{
\dontrun{
y <- t(MARSS::harborSealWA[, c("SJF", "SJI", "EBays")])
y[,10] <- y[,10] + 2
set.seed(1)
m <- fit_dfa(y = y, num_trends = 2, iter = 1000, chains = 1, nu_fixed = 2)
r <- rotate_trends(m)
p <- plot_trends(r)
print(p)
find_swans(r)
}
}
