% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_flipped_chains.R
\name{flip_chains}
\alias{flip_chains}
\title{Flip chains}
\usage{
flip_chains(model, trends = 1, print = FALSE, ...)
}
\arguments{
\item{model}{A Stan model.}

\item{trends}{The number of trends in the DFA}

\item{...}{Other arguments to pass to \code{\link{find_flipped_chains}}.}
}
\description{
Flip chains
}
