% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_looic.R
\name{get_looic}
\alias{get_looic}
\title{Get LOOIC}
\usage{
get_looic(fitted_model)
}
\arguments{
\item{fitted_model}{Output from \code{\link[=fit_dfa]{fit_dfa()}}.}
}
\description{
Extract the LOOIC (leave-one-out information criterion) using
\code{\link[loo:loo]{loo::loo()}}.
}
