% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dfa.R
\name{get_stan_model}
\alias{get_stan_model}
\title{Load compiled Stan model}
\usage{
get_stan_model(model_name)
}
\arguments{
\item{model_name}{String that is dfa, regime_1, hmm_gaussian, or corr}
}
\value{
Stan model.
}
\description{
Load compiled Stan model
}
