% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_chains.R
\name{invert_chains}
\alias{invert_chains}
\title{Invert chains}
\usage{
invert_chains(model, trends = 1, print = FALSE, ...)
}
\arguments{
\item{model}{A Stan model.}

\item{trends}{The number of trends in the DFA.}

\item{print}{Logical indicating whether the summary should be printed.}

\item{...}{Other arguments to pass to \code{\link{find_flipped_chains}}.}
}
\description{
Invert chains
}

