% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converge_rhat.R
\name{is_converged}
\alias{is_converged}
\title{Summarize Rhat convergence statistics across parameters}
\usage{
is_converged(fitted_model, threshold = 1.05)
}
\arguments{
\item{fitted_model}{A Stan model. E.g. a model from \code{\link{fit_dfa}}.}

\item{threshold}{Threshold for maximum Rhat, default = 1.05}
}
\description{
Pass in rstanfit model object, and optional threshold value for convergence. Returns boolean
}

