% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.bayesdfa}
\alias{loo.bayesdfa}
\title{Return LOO information criteria}
\usage{
\method{loo}{bayesdfa}(fitted_model, cores = getOption("mc.cores", 1))
}
\arguments{
\item{fitted_model}{Output from \code{\link[=fit_dfa]{fit_dfa()}}.}

\item{cores}{Number of cores to use for parallelization.}
}
\description{
Extract the LOOIC (leave-one-out information criterion) using
\code{\link[loo:loo]{loo::loo()}}.
}
\examples{
\donttest{
set.seed(1)
s <- sim_dfa(num_trends = 1, num_years = 50, num_ts = 3)
m <- fit_dfa(y = s$y_sim, iter = 300, chains = 1, num_trends = 1)
loo(m)
}
}
