% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loadings.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot the loadings from a DFA}
\usage{
plot_loadings(rotated_modelfit, names = NULL, threshold = 0.1,
  facet = TRUE)
}
\arguments{
\item{rotated_modelfit}{Output from \code{\link{rotate_trends}}.}

\item{names}{An optional vector of names for plotting the loadings.}

\item{threshold}{Threshold above which to plot the loadings.}

\item{facet}{Logical. Should there be a separate facet for each trend?}
}
\description{
Plot the loadings from a DFA
}

