% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_regimes2.R
\name{plot_regime_model}
\alias{plot_regime_model}
\title{Plot the state probabilities from \code{find_regimes2}}
\usage{
plot_regime_model(model, probs = c(0.1, 0.9))
}
\arguments{
\item{model}{A model returned by \code{\link{find_regimes2}}.}

\item{probs}{A numeric vector of quantiles to plot the credible intervals at.}
}
\description{
Plot the state probabilities from \code{find_regimes2}
}
\details{
Note that the original timeseries data (dots) are shown scaled between 0 and 1.
}
\examples{
\dontrun{
data(Nile)
m <- find_regimes2(log(Nile))
plot_regime_model(m)
}
}
