% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted.R
\name{predicted}
\alias{predicted}
\title{Calculate predicted value from DFA object}
\usage{
predicted(fitted_model)
}
\arguments{
\item{fitted_model}{Samples extracted (with \code{permuted = FALSE}) from a Stan
model. E.g. output from \code{\link[=invert_chains]{invert_chains()}}.}
}
\description{
Pass in \code{rstanfit} model object, and a threshold Rhat value for
convergence. Returns boolean.
}
