% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_cor.R
\name{trend_cor}
\alias{trend_cor}
\title{Estimate the correlation between a DFA trend and some other timeseries}
\usage{
trend_cor(rotated_modelfit, y, trend = 1, trend_samples = 100,
  stan_iter = 300, stan_chains = 1, ...)
}
\arguments{
\item{rotated_modelfit}{Output from \code{\link{rotate_trends}}}

\item{y}{A numeric vector to correlate with the DFA trend. Must be the same
length as the DFA trend.}

\item{trend}{A number corresponding to which trend to use.}

\item{trend_samples}{The number of samples from the trend posterior to use. A
model will be run for each trend sample so this value shouldn't be too
large.}

\item{stan_iter}{The number of samples from the posterior with each Stan
model run.}

\item{stan_chains}{The number of chains for each Stan model run.}

\item{...}{Other arguments to pass to \code{\link[rstan]{sampling}}}
}
\value{
A numeric vector of samples from the correlation coefficient posterior.
}
\description{
Fully incorporates the uncertainty from the posterior of the DFA trend
}
\examples{
\dontrun{
y <- t(MARSS::harborSealWA[, c("SJF", "SJI", "EBays", "PSnd")])
m <- fit_dfa(y = y, num_trends = 1, iter = 600, nu_fixed = 100)
r <- rotate_trends(m)
correlation <- trend_cor(r, y = rnorm(ncol(r$trends[,1,]), 0, 1))
hist(correlation)
}
}
