% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dfa_trends.R
\name{find_dfa_trends}
\alias{find_dfa_trends}
\title{Find the best number of trends according to LOO}
\usage{
find_dfa_trends(y = y, kmin = 1, kmax = 5, iter = 2000)
}
\arguments{
\item{y}{A matrix of data to fit. Columns represent time element.}

\item{kmin}{Minimum number of trends.}

\item{kmax}{Maximum number of trans.}

\item{iter}{Iterations when sampling from each Stan model.}
}
\description{
Fit a DFA with different number of trends and return the leave one out (LOO)
value as calculated by the loo package. Note that the models will be fit once
with shared variances and once with independent variances.
}

