% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dfa_trends.R
\name{find_dfa_trends}
\alias{find_dfa_trends}
\title{Find the best number of trends according to LOOIC}
\usage{
find_dfa_trends(y = y, kmin = 1, kmax = 5, iter = 2000,
  compare_normal = FALSE, convergence_threshold = 1.05,
  variance = c("equal", "unequal"), ...)
}
\arguments{
\item{y}{A matrix of data to fit. Columns represent time element.}

\item{kmin}{Minimum number of trends.}

\item{kmax}{Maximum number of trans.}

\item{iter}{Iterations when sampling from each Stan model.}

\item{compare_normal}{If TRUE, does model selection comparison of Normal vs
Student-t errors}

\item{convergence_threshold}{The maximum allowed value of Rhat to determine
convergence of parameters}

\item{variance}{Vector of variance arguments for searching over large groups
of models. Can be either or both of ("equal","unequal")}

\item{...}{Other arguments to pass to \code{\link{fit_dfa}}}
}
\description{
Fit a DFA with different number of trends and return the leave one out (LOO)
value as calculated by the loo package.
}
\examples{
\dontrun{
y <- t(MARSS::harborSealWA[, c("SJF", "SJI", "EBays")])
set.seed(1)
m <- find_dfa_trends(y = y, kmin = 1, kmax = 2,
  iter = 1000, chains = 1)
}
}
