% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_chains.R
\name{find_inverted_chains}
\alias{find_inverted_chains}
\title{Find which chains to invert}
\usage{
find_inverted_chains(model, trend = 1, thresh = 0.4, plot = FALSE)
}
\arguments{
\item{model}{A Stan model}

\item{trend}{Which trend to check}

\item{thresh}{A threshold for the required absolute difference in mean
standard deviation between the best permutation and the next best
permutation.}

\item{plot}{Should a plot of the trend for each chain be made?}
}
\description{
Finds which chains to invert by finding the best combination of chains to
multiply by -1 to minimize the standard deviation of the trends across the
chains.
}
\examples{
\dontrun{
y <- t(MARSS::harborSealWA[, c("SJF", "SJI", "EBays", "PSnd")])
set.seed(1)
m <- fit_dfa(y = y, num_trends = 2, iter = 2000, chains = 4)
# chains were already inverted, but we can redo that, as an example, with:
find_inverted_chains(m$model, trend = 1, plot = TRUE)
}
}
\seealso{
invert_chains
}
