% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dfa.R
\name{fit_dfa}
\alias{fit_dfa}
\title{Fit a Bayesian DFA}
\usage{
fit_dfa(y = y, covar = NULL, covar_index = NULL, num_trends = 2,
  varIndx = NULL, zscore = TRUE, iter = 4000, chains = 1,
  control = list(adapt_delta = 0.99), nu_fixed = 7, tau = 0.1,
  timevarying = FALSE, estimate_nu = FALSE)
}
\arguments{
\item{y}{A matrix of data to fit. Columns represent time element.}

\item{covar}{A matrix of covariates}

\item{covar_index}{Indices assigning ??}

\item{num_trends}{Number of trends to fit.}

\item{varIndx}{Indices indicating which timeseries should have shared variances.}

\item{zscore}{Logical. Should the data be standardized first?}

\item{iter}{Number of iterations in Stan sampling.}

\item{chains}{Number of chains in Stan sampling.}

\item{control}{A list of options to pass to Stan sampling.}

\item{nu_fixed}{Student t degrees of freedom parameter}

\item{tau}{A fixed parameter describing the standard deviation on the random
walk for the factor loadings in the case of time varying DFA.}

\item{timevarying}{Logical. If \code{TRUE}, a time varying DFA. Note that the
time varying DFA has not been extensively tested and may not return
sensible answers.}
}
\description{
Fit a Bayesian DFA
}

