% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dfa.R
\name{fit_dfa}
\alias{fit_dfa}
\title{Fit a Bayesian DFA}
\usage{
fit_dfa(y = y, covar = NULL, covar_index = NULL, num_trends = 2,
  varIndx = NULL, zscore = TRUE, iter = 2000, chains = 4,
  control = list(adapt_delta = 0.99), nu_fixed = 7, tau = 0.1,
  timevarying = FALSE, estimate_nu = FALSE)
}
\arguments{
\item{y}{A matrix of data to fit. Columns represent time element.}

\item{covar}{A matrix of covariates}

\item{covar_index}{Indices assigning ??}

\item{num_trends}{Number of trends to fit.}

\item{varIndx}{Indices indicating which timeseries should have shared
variances.}

\item{zscore}{Logical. Should the data be standardized first?}

\item{iter}{Number of iterations in Stan sampling.}

\item{chains}{Number of chains in Stan sampling.}

\item{control}{A list of options to pass to Stan sampling.}

\item{nu_fixed}{Student t degrees of freedom parameter}

\item{tau}{A fixed parameter describing the standard deviation on the random
walk for the factor loadings in the case of time varying DFA.}

\item{timevarying}{Logical. If \code{TRUE}, a time varying DFA. Note that the
time varying DFA has not been extensively tested and may not return
sensible answers.}

\item{estimate_nu}{Logical. Estimate the student t degrees of freedom
parameter?}
}
\description{
Fit a Bayesian DFA
}
\details{
Note that there is nothing restricting the loadings and trends from
  being inverted (multiplied by -1) for a given chain. Therefore, if you fit
  multiple chains, the package will attempt to determine which chains need to
  be inverted using the function \code{\link{find_inverted_chains}}.
}
\references{
Del Negro, M., & Otrok, C. (2008). Dynamic factor models with time-varying
parameters: measuring changes in international business cycles.
}
\seealso{
plot_loadings plot_trends rotate_trends
}
