% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_chains.R
\name{invert_chains}
\alias{invert_chains}
\title{Invert chains}
\usage{
invert_chains(model, trends = 1, print = FALSE, ...)
}
\arguments{
\item{model}{A Stan model, rstanfit object}

\item{trends}{The number of trends in the DFA, defaults to 1}

\item{print}{Logical indicating whether the summary should be printed.
Defaults to \code{FALSE}.}

\item{...}{Other arguments to pass to \code{\link[=find_inverted_chains]{find_inverted_chains()}}.}
}
\description{
Invert chains
}
\seealso{
find_inverted_chains
}
