% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted.R
\name{plot_fitted}
\alias{plot_fitted}
\title{Plot the trends from a DFA}
\usage{
plot_fitted(modelfit, names = NULL)
}
\arguments{
\item{modelfit}{Output from \code{\link{fit_dfa}}, a rstanfit object}

\item{names}{Optional vector of names for plotting labels}
}
\description{
Plot the trends from a DFA
}
\examples{
y <- t(MARSS::harborSealWA[, c("SJF", "SJI", "EBays", "PSnd")])
m <- fit_dfa(y = y, num_trends = 2, iter = 250, chains = 1)
p <- plot_fitted(m)
print(p)
}
\seealso{
plot_loadings fit_dfa rotate_trends
}
