% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loadings.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot the loadings from a DFA}
\usage{
plot_loadings(rotated_modelfit, names = NULL, facet = TRUE, violin = TRUE,
  conf_level = 0.95)
}
\arguments{
\item{rotated_modelfit}{Output from \code{\link{rotate_trends}}.}

\item{names}{An optional vector of names for plotting the loadings.}

\item{facet}{Logical. Should there be a separate facet for each trend?}

\item{violin}{Logical. Should the full posterior densities be shown as a
violin plot?}

\item{conf_level}{Confidence level for credible intervals.}
}
\description{
Plot the loadings from a DFA
}
\examples{
y <- t(MARSS::harborSealWA[, c("SJF", "SJI", "EBays", "PSnd")])
m <- fit_dfa(y = y, num_trends = 2, iter = 800)
r <- rotate_trends(m)
plot_loadings(r, violin = FALSE, facet = TRUE)
plot_loadings(r, violin = FALSE, facet = FALSE)
plot_loadings(r, violin = TRUE, facet = FALSE)
plot_loadings(r, violin = TRUE, facet = TRUE)
}
\seealso{
plot_trends fit_dfa rotate_trends
}

