% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loadings.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot the loadings from a DFA}
\usage{
plot_loadings(rotated_modelfit, names = NULL, threshold = 0.8,
  facet = TRUE)
}
\arguments{
\item{rotated_modelfit}{Output from \code{\link{rotate_trends}}.}

\item{names}{An optional vector of names for plotting the loadings.}

\item{threshold}{Numeric (0-1). Optional, if included, only plot loadings who have Pr(<0) or Pr(>0) > threshold}

\item{facet}{Logical. Should there be a separate facet for each trend?}
}
\description{
Plot the loadings from a DFA
}
\examples{
y <- t(MARSS::harborSealWA[, c("SJF", "SJI", "EBays", "PSnd")])
m <- fit_dfa(y = y, num_trends = 2, iter = 500)
r <- rotate_trends(m)
p <- plot_loadings(r, threshold = 0.01, facet = FALSE)
print(p)
}
\references{
Del Negro, M., & Otrok, C. (2008). Dynamic factor models with time-varying
parameters: measuring changes in international business cycles.
}

