% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trends.R
\name{plot_trends}
\alias{plot_trends}
\title{Plot the trends from a DFA}
\usage{
plot_trends(rotated_modelfit)
}
\arguments{
\item{rotated_modelfit}{Output from \code{\link{rotate_trends}}}
}
\description{
Plot the trends from a DFA
}

