% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_trends.R
\name{rotate_trends}
\alias{rotate_trends}
\title{Rotate the trends from a DFA}
\usage{
rotate_trends(fitted_model, conf_level = 0.95)
}
\arguments{
\item{fitted_model}{Output from \code{\link{fit_dfa}}.}

\item{conf_level}{Probability level for CI.}
}
\description{
Rotate the trends from a DFA
}

