% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_trends.R
\name{rotate_trends}
\alias{rotate_trends}
\title{Rotate the trends from a DFA}
\usage{
rotate_trends(fitted_model)
}
\arguments{
\item{fitted_model}{Output from \code{\link{fit_dfa}}}
}
\description{
Rotate the trends from a DFA
}

